requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
pyautogen>=0.2.0
openai>=1.66.2
ag2[openai] 
pyyaml
pinecone
langchain
langchain_openai